refactor: remove `OrderedBatchProducerV1`
