+++ src/log-store/src/kafka.rs
@@ -14,11 +14,6 @@
 
 pub(crate) mod client_manager;
 pub mod log_store;
-// TODO(weny): remove it.
-#[allow(dead_code)]
-pub(crate) mod background_producer;
-// TODO(weny): remove it.
-#[allow(dead_code)]
 pub(crate) mod producer;
 pub(crate) mod producer_registry;
 pub(crate) mod util;
